EMANUELE -metodo getAvg_Nsegnalazioni in ServerCV.java

+++ src/main/java/org/example/centrivaccinali/gui/Client03CT_CercaCV_Controller.java
@@ -158,7 +158,7 @@ public class Client03CT_CercaCV_Controller implements Initializable {
         }
 
         //calcolo numero segnalazioni
-        nSegnalazioni.setText(String.valueOf(mediaSeveritaEventiAvversi[1]));
+        nSegnalazioni.setText(String.valueOf((int)mediaSeveritaEventiAvversi[1]));
 
         //impostazione grafica
         chosenCV.setStyle("-fx-background-radius: 30;");
